# openapi.yaml
openapi: 3.0.0
info:
  x-skaffolder-id-project: 63307f5ac97c70218d65d9e0
  title: Zuct
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
  /login:
    post:
      x-skaffolder-ignore: true
      summary: Login API
      security: []
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              example:
                username: admin
                password: 62f264d7ad826f02a8af714c0a54b197935b717656b80461686d450f7b3abde4c553541515de2052b9af70f710f0cd8a1a2d3f4d60aa72608d71a63a9a93c0f5
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /verifyToken:
    post:
      x-skaffolder-ignore: true
      summary: Verify token API
      security: []
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /changePassword:
    post:
      x-skaffolder-ignore: true
      summary: Change password API
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                passwordNew:
                  type: string
                passwordOld:
                  type: string
      responses:
        "200":
          description: OK
...
  /user:
    post:
      x-skaffolder-id: 63307f5bc97c70218d65d9f2
      x-skaffolder-name: create
      x-skaffolder-id-resource: 63307f5ac97c70218d65d9ea
      x-skaffolder-resource: User
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mail:
                  type: string
                name:
                  type: string
                password:
                  type: string
                roles:
                  type: string
                surname:
                  type: string
                username:
                  type: string
    get:
      x-skaffolder-id: 63307f5bc97c70218d65d9ed
      x-skaffolder-name: list
      x-skaffolder-id-resource: 63307f5ac97c70218d65d9ea
      x-skaffolder-resource: User
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  "/user/{id}":
    delete:
      x-skaffolder-id: 63307f5bc97c70218d65d9ee
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 63307f5ac97c70218d65d9ea
      x-skaffolder-resource: User
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 63307f5bc97c70218d65d9f3
      x-skaffolder-name: get
      x-skaffolder-id-resource: 63307f5ac97c70218d65d9ea
      x-skaffolder-resource: User
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    post:
      x-skaffolder-id: 63307f5bc97c70218d65d9f0
      x-skaffolder-name: update
      x-skaffolder-id-resource: 63307f5ac97c70218d65d9ea
      x-skaffolder-resource: User
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
  /courses:
    post:
      x-skaffolder-id: 63308788c97c70218d65da80
      x-skaffolder-name: create
      x-skaffolder-id-resource: 63308546c97c70218d65da54
      x-skaffolder-resource: courses
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - courses
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/courses"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                course_code:
                  type: object
                course_name:
                  type: string
    get:
      x-skaffolder-id: 63308788c97c70218d65da84
      x-skaffolder-name: list
      x-skaffolder-id-resource: 63308546c97c70218d65da54
      x-skaffolder-resource: courses
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - courses
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/courses"
  "/courses/{id}":
    delete:
      x-skaffolder-id: 63308788c97c70218d65da85
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 63308546c97c70218d65da54
      x-skaffolder-resource: courses
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - courses
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 63308788c97c70218d65da81
      x-skaffolder-name: get
      x-skaffolder-id-resource: 63308546c97c70218d65da54
      x-skaffolder-resource: courses
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - courses
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/courses"
    post:
      x-skaffolder-id: 63308788c97c70218d65da7e
      x-skaffolder-name: update
      x-skaffolder-id-resource: 63308546c97c70218d65da54
      x-skaffolder-resource: courses
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - courses
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/courses"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/courses"
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      x-skaffolder-id: 63307f5ac97c70218d65d9ea
      x-skaffolder-id-db: 63307f5ac97c70218d65d9e1
      x-skaffolder-id-entity: 63307f5ac97c70218d65d9e2
      x-skaffolder-url: /user
      x-skaffolder-type: User
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        mail:
          type: string
          x-skaffolder-id-attr: 63307f5ac97c70218d65d9e5
          x-skaffolder-type: String
        name:
          type: string
          x-skaffolder-id-attr: 63307f5ac97c70218d65d9e6
          x-skaffolder-type: String
        password:
          type: string
          x-skaffolder-id-attr: 63307f5ac97c70218d65d9e4
          x-skaffolder-type: String
          x-skaffolder-required: true
        roles:
          type: string
          x-skaffolder-id-attr: 63307f5ac97c70218d65d9e8
          x-skaffolder-type: String
        surname:
          type: string
          x-skaffolder-id-attr: 63307f5ac97c70218d65d9e7
          x-skaffolder-type: String
        username:
          type: string
          x-skaffolder-id-attr: 63307f5ac97c70218d65d9e3
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations: 
      required:
        - _id
        - password
        - username
    courses:
      x-skaffolder-id: 63308546c97c70218d65da54
      x-skaffolder-id-db: 63307f5ac97c70218d65d9e1
      x-skaffolder-id-entity: 63308546c97c70218d65da51
      x-skaffolder-url: /courses
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        course_code:
          type: object
          x-skaffolder-id-attr: 63308546c97c70218d65da52
          x-skaffolder-type: Custom
          x-skaffolder-required: true
        course_name:
          type: string
          x-skaffolder-id-attr: 63308546c97c70218d65da53
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations:
        _contains:
          x-skaffolder-id: 63308b4ac97c70218d65da96
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 63308546c97c70218d65da51
          x-skaffolder-ent2: 633086dfc97c70218d65da64
        _teaches:
          x-skaffolder-id: 63308b4ac97c70218d65da97
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 63308546c97c70218d65da51
          x-skaffolder-ent2: 633085b7c97c70218d65da59
      required:
        - _id
        - course_code
        - course_name
...